Добавляет ДЗ Модуляция
